# Closet UI Update
## Default view
Default view is going to be moved to the current edit mode, to allow the user to view more of their clothes easily.

## Image click function
Pressing on an image in the edit mode now will not prompt the user to delete the image

## UI difference
Now when an image is pressed a selector will show at the bottom, with the image type. i.e. Gym piece
followed by the add to favourites(ADDED LATER), add to fit(ADDED LATER), and the relocated delete button.