# Login update
## User awareness
When the user signed in previously the user wasn't let now of the login
status. The app should notify the user about whether their sign in was succesful or not,

## Password change
Previously a change of password was to be requested to the ADMIN in order to send a change of password email.
Now this is done by the USER themself and no admin intervention required anymore via a link on the login page.